all works butlarge tests

+++ src/common.h
@@ -41,7 +41,7 @@ typedef uint64_t u64;
 
 #define DBC_CHECK_VID(vid) DBC_CHECK(vid > 0 && vid < MAX_SET_SIZE, "VariableId is out of range")
 
-typedef u32 InstanceId;
+typedef u64 InstanceId;
 typedef unsigned int VarId;
 typedef float ValueType;
 // Value of variable within its domain definition
